requirements.txt
paho-mqtt==1.6.1
watchdog==3.0.0
pyyaml==6.0.1
requests==2.31.0 